refactor: 기본 enum클래스들의 패키지 위치 조정
