+++ src/main/java/lotto/io/MessagePrinter.java
@@ -4,19 +4,19 @@ import lotto.domain.Lotto;
 import java.util.List;
 import java.util.Map;
 
-import static lotto.LottoConstants.LOTTO_PRICE;
-import static lotto.OutputMessages.MONEY_INPUT_REQUEST_MESSAGE;
-import static lotto.OutputMessages.WINNING_NUMBERS_INPUT_REQUEST_MESSAGE;
-import static lotto.OutputMessages.BONUS_NUMBER_INPUT_REQUEST_MESSAGE;
-import static lotto.OutputMessages.LOTTO_PURCHASE_MESSAGE;
-import static lotto.OutputMessages.WINNING_STATISTICS_MESSAGE;
-import static lotto.OutputMessages.DIVIDING_LINE;
-import static lotto.OutputMessages.EARNING_RATE_MESSAGE;
-import static lotto.OutputMessages.FIFTH_STATISTICS_MESSAGE;
-import static lotto.OutputMessages.FOURTH_STATISTICS_MESSAGE;
-import static lotto.OutputMessages.THIRD_STATISTICS_MESSAGE;
-import static lotto.OutputMessages.SECOND_STATISTICS_MESSAGE;
-import static lotto.OutputMessages.FIRST_STATISTICS_MESSAGE;
+import static lotto.domain.LottoConstants.LOTTO_PRICE;
+import static lotto.io.OutputMessages.MONEY_INPUT_REQUEST_MESSAGE;
+import static lotto.io.OutputMessages.WINNING_NUMBERS_INPUT_REQUEST_MESSAGE;
+import static lotto.io.OutputMessages.BONUS_NUMBER_INPUT_REQUEST_MESSAGE;
+import static lotto.io.OutputMessages.LOTTO_PURCHASE_MESSAGE;
+import static lotto.io.OutputMessages.WINNING_STATISTICS_MESSAGE;
+import static lotto.io.OutputMessages.DIVIDING_LINE;
+import static lotto.io.OutputMessages.EARNING_RATE_MESSAGE;
+import static lotto.io.OutputMessages.FIFTH_STATISTICS_MESSAGE;
+import static lotto.io.OutputMessages.FOURTH_STATISTICS_MESSAGE;
+import static lotto.io.OutputMessages.THIRD_STATISTICS_MESSAGE;
+import static lotto.io.OutputMessages.SECOND_STATISTICS_MESSAGE;
+import static lotto.io.OutputMessages.FIRST_STATISTICS_MESSAGE;
 import static lotto.domain.Rank.FIFTH;
 import static lotto.domain.Rank.FOURTH;
 import static lotto.domain.Rank.THIRD;
