+++ src/test/java/lotto/domain/generator/LottoGameResultGeneratorTest.java
@@ -1,0 +1,2 @@
+package lotto.domain.generator;public class LottoGameResultGeneratorTest {
+}
